requirements.txt
fastapi
uvicorn
autogluon.tabular
pandas
